requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
boto3==1.18.31
botocore==1.21.31
dj-database-url==0.5.0
django==3.2.6
django-filter==2.4.0
django-redis==5.0.0
django-storages==1.11.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
haikunator==2.1.0
jmespath==0.10.0
promise==2.3
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
rx==1.6.1
s3transfer==0.5.0
setuptools==57.4.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0; python_version < "3.8"
urllib3==1.26.6
whitenoise==5.3.0
